[us40] Fix baseFolderService not used anywhere but in resolver causing unexpected issue
